// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
// ...
    Functions: {
// ...
      get_enriched_topics: {
        Args: { p_category_id?: string; p_limit?: number; p_offset?: number }
        Returns: {
          id: string
          title: string
          content: string
          author_id: string
          category_id: string
          is_pinned: boolean
          is_locked: boolean
          view_count: number
          reply_count: number
          last_reply_at: string
          created_at: string
          updated_at: string
          slug: string
          moderation_status: string
          last_post_id: string
          author_username: string
          author_avatar_url: string
          category_name: string
          category_color: string
          category_slug: string
          parent_category_id: string
        }[]
      }
// ...
    }
    Enums: {
      user_role: "admin" | "moderator" | "user"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}